address review comments
