+++ src/speculators/models/dflash/attention.py
@@ -1,11 +1,8 @@
 # ruff: noqa: ERA001
 import torch
 from torch.nn.attention.flex_attention import (
-    BlockMask,
     or_masks,
 )
-
-from speculators.models.attention import ALL_ATTENTION_FUNCTIONS  # noqa: F401
 
 
 def create_anchor_block_mask_mod(
@@ -117,18 +114,3 @@ def create_anchor_block_mask_mod(
         return kv_is_block & (q_block == kv_block)
 
     return or_masks(base_prefix_mod, same_block_mod), q_len, kv_len
-
-
-def block_mask_to_dense_attention_mask(
-    block_mask: BlockMask, device: torch.device, dtype: torch.dtype
-):
-    attention_mask = torch.ones(block_mask.shape, device=device, dtype=dtype)
-
-    for q_idx in range(attention_mask.shape[2]):
-        attention_mask[0, 0, q_idx, :] = block_mask.mask_mod(
-            torch.zeros(1, device=device, dtype=torch.long),
-            torch.zeros(1, device=device, dtype=torch.long),
-            torch.ones(1, device=device, dtype=torch.long) * q_idx,
-            torch.arange(attention_mask.shape[3], device=device, dtype=torch.long),
-        )
-    return attention_mask
